move NoDiskConflict predicate to its filter plugin
